fix warnings

+++ src/Topk.cc
@@ -178,7 +178,7 @@ bool TopkVal::DoSerialize(SerialInfo* info) const
 	else 
 		assert(numElements == 0);
 
-	int i = 0;
+	uint64_t i = 0;
 	std::list<Bucket*>::const_iterator it = buckets.begin();
 	while ( it != buckets.end() ) 
 		{
@@ -223,7 +223,7 @@ bool TopkVal::DoUnserialize(UnserialInfo* info)
 	else
 		assert(numElements == 0);
 
-	int i = 0;
+	uint64_t i = 0;
 	while ( i < numElements ) 
 		{
 		Bucket* b = new Bucket();
@@ -232,7 +232,7 @@ bool TopkVal::DoUnserialize(UnserialInfo* info)
 		v &= UNSERIALIZE(&b->count);
 		b->bucketPos = buckets.insert(buckets.end(), b);
 
-		for ( int j = 0; j < elements_count; j++ ) 
+		for ( uint64_t j = 0; j < elements_count; j++ ) 
 			{
 			Element* e = new Element();
 			v &= UNSERIALIZE(&e->epsilon);
